Move pil_to_tensor() and tensor_to_pil() utilities to the SpandrelImageToImage class.
